Move fast tests in own class
